+++ packages/ti/board/diag/current_monitor/src/current_monitor_test.h
@@ -114,7 +114,7 @@ extern "C" {
  *  \brief Structure defining INA Device calibration parameters.
  */
 typedef struct  CalibrationParams{
-    uint8_t maxRShunt;
+    float maxRShunt;
     float shuntVolLsb;
     float busVolLsb;
     float powerLsb;
